sirenia: Move trichechus over to EventMultiplexer.
This reworks trichechus to use EventMultiplexer to handle syslog,
incoming connections, and messages on a single thread. Further work is
needed to implement log message forwarding and app instance cleanup.

BUG=b:144915425
TEST=dbus-send --system --dest=org.chromium.ManaTEE --type=method_call\
  --print-reply /org/chromium/ManaTEE1 org.chromium.ManaTEEInterface.\
StartTEEApplication string:'shell'

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2490658
Commit-Queue: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>

+++ sirenia/src/communication/mod.rs
@@ -69,7 +69,7 @@ pub enum Response {
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct AppInfo {
     pub app_id: String,
-    pub port_number: u16,
+    pub port_number: u32,
 }
 
 pub fn get_app_path(id: &str) -> Result<&str> {
